+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/users/admin.py
@@ -1,10 +1,15 @@
+from django.conf import settings
 from django.contrib import admin
 from django.contrib.auth import admin as auth_admin
-from django.contrib.auth import get_user_model
+from django.contrib.auth import get_user_model, decorators
 
 from {{ cookiecutter.project_slug }}.users.forms import UserChangeForm, UserCreationForm
 
 User = get_user_model()
+
+if settings.DJANGO_ADMIN_FORCE_ALLAUTH:
+    # Force the `admin` sign in process to go through the `django-allauth` workflow
+    admin.site.login = decorators.login_required(admin.site.login)
 
 
 @admin.register(User)